requirements.txt
streamlit
python-pptx
fpdf
openai>=1.1.0
google-generativeai
pandas
reportlab





